requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.0
python-dotenv==0.19.2
flask-cors==4.0.1